concepts/initcalls: rewording and typo corrections
This patch handle different typos present in the initcall section of the book.
It also reword some paragraphs in order to get it clearer to the reader.
Some example codes were also added in order to help the reader to understand
smaller concepts, like the way initcalls callbacks are named in the object
file.

Signed-off-by: Bruno Meneguele <bmeneguele@gmail.com>