Initialization/linux-initialization-8.md: Fix grammatical errors & improve wording
s/as if a single task/as if they were a single task/
s/The least unit which scheduler operates is an individual task or thread. But a process is not only one type of entities of which the scheduller may operate./
The smallest unit that the scheduler works with is an individual task or thread. However, a process is not the only type of entity that the scheduler can operate with./
s/these options provides support/these options provide support/
s/The first one option provides support for group scheduling with `completely fair scheduler` policies and the second with `real-time` policies respectively./
The first option provides support for group scheduling with the `completely fair scheduler` policies and the second with the `real-time` policies respectively./
s/It means that it supports/That means it supports/