linux-interrupts-2: Improve text block
Remove redundant 'the'
s/Now as we jumped on the startup_32 from the/Now as we jumped into `startup_32` from/

Remove redundant 'the', ensure a varied choice of words
s/we will not find anything related to the interrupt handling here/
  we will not encounter anything related to interrupt handling here/

Improve sentence structure
s/The startup_32 contains code that makes preparations before the transition into long mode and directly jumps in to it/
The code inside of `startup_32` makes necessary preparations, before transitioning into the long mode with a direct jump/

Improve sentence structure and ensure a varied choice of words.
s/it makes preparations before the kernel decompression that occurs in the decompress_kernel from the/
it makes arrangements for the kernel decompression that occurs in the `decompress_kernel` function inside of ../

Remove redundant 'the', vary the choice of words
s/we jump on the startup_64 from the/we jump into `startup_64` defined at/

Reduce repetition, improve sentence structure
s/we start to build identity-mapped pages. After we have built identity-mapped pages, checked the NX bit, setup the Extended Feature Enable Register (see in links), and updated the early Global Descriptor Table with the lgdt instruction, we need to setup gs register/

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>