Create a release with ebooks on push to master
Use GitHub actions to create a release with PDF, EPUB, and MOBI files
automatically on push to master. The action can also be triggered
manually.

The commit ID will be used as tag for releases.

This action does not use GitBook, but a fork of it named HonKit.
gitbook-cli isn't maintained since 6 years, whereas honkit is a
maintained fork.