Generate e-books for every release and PR
There are still some code blocks which do not fit the A5 format. Let's make A4-sized PDFs the default size. The A5 PDF may still be the oly format suitable for e-book readers, though, as MOBI and EPUB mess code blocks expected to be written in a monotype font up.

Bitmap images are still somewhat oversized in the A4 format, though. A CSS rule similar to the following one may be needed in a CSS file at the right location in the repository:

@media print {
    img {
        object-fit: scale-down;
    }
}